requirements.txt
numpy ~= 1.19
wpilib ~= 2021.2.1
robotpy-ctre ~= 2021.1.2
robotpy-navx ~= 2021.0.0
robotpy-rev ~= 2021.0.0
robotpy-wpilib-utilities ~= 2021.0.1
robotpy-wpimath ~= 2021.2.1.2
